[IDLE-300] 코드 정리
